Changes requested by @jreback
