+++ pandas/tests/extension/arrow/test_bool.py
@@ -39,14 +39,6 @@ class TestInterface(BaseArrowTests, base.BaseInterfaceTests):
     def test_repr(self, data):
         raise pytest.skip("TODO")
 
-    def test_shift_non_empty_array(self, data):
-        raise pytest.skip("Skipping because the implementation of"
-                          "`ArrowBoolArray` is quite buggy.")
-
-    def test_shift_empty_array(self, data):
-        raise pytest.skip("Skipping because the implementation of"
-                          "`ArrowBoolArray` is quite buggy.")
-
 
 class TestConstructors(BaseArrowTests, base.BaseConstructorsTests):
     def test_from_dtype(self, data):
