+++ pandas/tests/extension/base/interface.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 
 from pandas.compat import StringIO
 
@@ -87,25 +86,3 @@ class BaseInterfaceTests(BaseExtensionTests):
             assert not na.all()
 
             assert na.dtype._is_boolean
-
-    @pytest.mark.parametrize('periods, indices', [
-        [-4, [-1, -1]],
-        [-1, [1, -1]],
-        [0, [0, 1]],
-        [1, [-1, 0]],
-        [4, [-1, -1]]
-    ])
-    def test_shift_non_empty_array(self, data, periods, indices):
-        subset = data[:2]
-        result = subset.shift(periods)
-        expected = subset.take(indices, allow_fill=True)
-        self.assert_extension_array_equal(result, expected)
-
-    @pytest.mark.parametrize('periods', [
-        -4, -1, 0, 1, 4
-    ])
-    def test_shift_empty_array(self, data, periods):
-        empty = data[:0]
-        result = empty.shift(periods)
-        expected = empty
-        self.assert_extension_array_equal(result, expected)
